pyproject.toml
[project]
name = "whatsapp-teachers-to-sheets"
version = "1.0.0"
description = "Слушает WhatsApp-чат родителей, извлекает номера учителей через Gemini, пишет в Google Таблицы. MCP-сервер для управления."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-storage>=2.14.0",
    "pywa[fastapi]>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.0",
    "bcrypt>=4.0.0",
    "itsdangerous>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
